+++ src/main/java/tech/habegger/elastic/shared/GeoCoord.java
@@ -1,7 +1,7 @@
 package tech.habegger.elastic.shared;
 
-public record GeoCoord(float lat, float lon) {
-    public static GeoCoord geoCoord(float lat, float lon) {
+public record GeoCoord(double lat, double lon) {
+    public static GeoCoord geoCoord(double lat, double lon) {
         return new GeoCoord(lat, lon);
     }
 }
